Migration fix in Version20190403120728: check for existing columns, see also #5947

+++ bundles/CoreBundle/Resources/migrations/Version20190403120728.php
@@ -34,17 +34,18 @@ class Version20190403120728 extends AbstractPimcoreMigration
                 $metaTable = current($table);
 
                 // add index column to metadata tables
-                $this->addSql('ALTER TABLE `' . $metaTable . '` ADD COLUMN `index` int(11) unsigned NOT NULL DEFAULT \'0\'');
-                $this->addSql('ALTER TABLE `' . $metaTable . '` DROP PRIMARY KEY');
-                $this->addSql('ALTER TABLE `' . $metaTable . '` ADD PRIMARY KEY (`o_id`, `dest_id`, `type`, `fieldname`, `column`, `ownertype`, `ownername`, `position`, `index`)');
+                if(!$schema->getTable($metaTable)->hasColumn('index')) {
+                    $this->addSql('ALTER TABLE `' . $metaTable . '` ADD COLUMN `index` int(11) unsigned NOT NULL DEFAULT \'0\'');
+                    $this->addSql('ALTER TABLE `' . $metaTable . '` DROP PRIMARY KEY');
+                    $this->addSql('ALTER TABLE `' . $metaTable . '` ADD PRIMARY KEY (`o_id`, `dest_id`, `type`, `fieldname`, `column`, `ownertype`, `ownername`, `position`, `index`)');
 
-                //update index in metadata tables from relations tables
-                $relTable = explode('_', $metaTable);
-                $relTable = 'object_relations_' . $relTable[2];
+                    //update index in metadata tables from relations tables
+                    $relTable = explode('_', $metaTable);
+                    $relTable = 'object_relations_' . $relTable[2];
 
-                // Don't use INNER JOIN to update table as we could hita MySQL Issue with it.
-                // Details: https://github.com/pimcore/pimcore/issues/4233
-                $this->addSql('UPDATE `' . $metaTable . '` mt
+                    // Don't use INNER JOIN to update table as we could hita MySQL Issue with it.
+                    // Details: https://github.com/pimcore/pimcore/issues/4233
+                    $this->addSql('UPDATE `' . $metaTable . '` mt
                       SET mt.index = (SELECT rl.index FROM `' . $relTable . '` rl
                         WHERE  mt.o_id = rl.src_id
                             AND mt.dest_id = rl.dest_id
@@ -53,6 +54,7 @@ class Version20190403120728 extends AbstractPimcoreMigration
                             AND mt.ownertype = rl.ownertype
                             AND mt.ownername = rl.ownername
                             AND mt.position = rl.position)');
+                }
             }
         } catch (\Exception $e) {
             $this->writeMessage('An error occurred while performing migrations: ' . $e->getMessage());
@@ -78,9 +80,11 @@ class Version20190403120728 extends AbstractPimcoreMigration
             $metaTables = $db->fetchAll("SHOW TABLES LIKE 'object\_metadata\_%'");
             foreach ($metaTables as $table) {
                 $metaTable = current($table);
-                $this->addSql('ALTER TABLE `' . $metaTable . '` DROP COLUMN `index`');
-                $this->addSql('ALTER TABLE `' . $metaTable . '` DROP PRIMARY KEY');
-                $this->addSql('ALTER TABLE `' . $metaTable . '` ADD PRIMARY KEY (`o_id`, `dest_id`, `type`, `fieldname`, `column`, `ownertype`, `ownername`, `position`)');
+                if($schema->getTable($metaTable)->hasColumn('index')) {
+                    $this->addSql('ALTER TABLE `' . $metaTable . '` DROP COLUMN `index`');
+                    $this->addSql('ALTER TABLE `' . $metaTable . '` DROP PRIMARY KEY');
+                    $this->addSql('ALTER TABLE `' . $metaTable . '` ADD PRIMARY KEY (`o_id`, `dest_id`, `type`, `fieldname`, `column`, `ownertype`, `ownername`, `position`)');
+                }
             }
         } catch (\Exception $e) {
             $this->writeMessage('An error occurred while performing migrations: ' . $e->getMessage());
